Submit puids in renamealbum
Move PUID submission out of albumidentify so it can be run without
any side effects. Use the musicbrainz web api to submit PUIDs.

+++ albumidentify.py
@@ -12,7 +12,6 @@ import pickle
 import hashlib
 import random
 import shelve
-import puidsubmit
 import albumidentifyconfig
 import re
 import sets
@@ -113,11 +112,14 @@ def populate_fingerprint_cache(fname):
 		decode(fname,toname)
 		update_progress("Generating fingerprint")
 		(fp, dur) = fingerprint.fingerprint(toname)
+	except Exception, e:
+		print "Error creating fingerprint:",e
+		raise e
 	finally:
-		os.unlink(toname)
+		if os.path.exists(toname):
+			os.unlink(toname)
 
 	return fp, dur
-	
 
 def hash_file(fname):
 	update_progress("Hashing file")
@@ -384,55 +386,6 @@ def choose_track(possible_releases, track_generator, trackinfo):
 			return fileid
 	return fileid
 
-def all_tracks_in_order(release,trackinfo,trackmap):
-	file_ids = trackinfo.keys()
-	file_ids.sort()
-	for fileid in trackinfo:
-		found_tracknumber=lookups.track_number(release.tracks, track)
-		if found_tracknumber != file_ids.index(fileid)+1:
-			return False
-	return True
-
-def submit_shortcut_puids(releaseid,trackinfo,releaseinfo):
-	if not albumidentifyconfig.config.getboolean("albumidentify",
-		"push_shortcut_puids"):
-		print "Not submiting shortcut puids: not enabled in config"
-		return
-	release = lookups.get_release_by_releaseid(releaseid)
-	if not FORCE_ORDER:
-		print "Not submitting: No order"
-		return
-	flag=0
-	puid2trackid={}
-	for trackind in range(len(releaseinfo)):
-		trackid = release.tracks[trackind].id
-		puid = trackinfo[releaseinfo[trackind+1]][5]
-		if puid == None:
-			if not flag:
-				print "Submitting shortcut puids to musicbrainz"
-				print release.artist.name,"-",release.title,":"
-				flag=1
-			print "%02d" % (trackind+1),": No PUID"
-			continue
-		elif trackid not in [t.id for t in lookups.get_tracks_by_puid(puid)]:
-			if not flag:
-				print "Submitting shortcut puids to musicbrainz"
-				print release.artist.name,"-",release.title,":"
-				flag=1
-			print "%02d:" % (trackind+1),puid,"->",release.tracks[trackind].title
-			puid2trackid[puid]=trackid
-		elif puid not in lookups.get_track_by_id(trackid).puids:
-			# There is a mapping of puid -> trackid, but not of trackid->puid,
-			# this almost certainly means our cache is out of date, so delete it.
-			# This just meant that searching was slower, so next time we can find this
-			# faster
-			lookups.remove_from_cache("delayed_get_track_by_id",trackid)
-	if flag:
-		print
-		puidsubmit.submit_puids(puid2trackid)
-	else:
-		print "No shortcut puids need submitting"
-
 def giving_up(removed_releases,fileid):
 	if removed_releases:
 		print "Possible releases:"
@@ -623,9 +576,6 @@ def guess_album2(trackinfo):
 					and releaseid not in completed_releases:
 				print release.title,"seems ok\x1b[K"
 				print "Musicbrainz Release Id:",release.id
-				submit_shortcut_puids(releaseid,
-					trackinfo,
-					possible_releases[releaseid])
 				yield releaseid, possible_releases[releaseid]
 				completed_releases.append(releaseid)
